backend/fastapi/requirements_fastapi.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.2
sqlalchemy>=1.4.23
asyncpg>=0.27.0
psycopg2-binary>=2.9.1
alembic>=1.7.3
python-dotenv>=0.19.0
passlib>=1.7.4
python-jose>=3.3.0
python-multipart>=0.0.5
email-validator>=1.1.3
starlette>=0.14.2
typing-extensions>=3.10.0
dependency-injector>=4.41.0
requests>=2.28.0
google-auth>=2.16.0
httpx>=0.24.0